requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0 
matplotlib>=3.7.0
seaborn>=0.12.0
# Transformers and model libraries
transformers>=4.35.0
torch>=2.0.0
peft>=0.6.0
datasets>=2.14.0
huggingface_hub>=0.17.0
numpy>=1.24.0
scikit-learn>=1.3.0

# System utilities
psutil>=5.9.0

# ================================
# API and Web Framework
# ================================

flask>=2.3.0

# ================================
# External AI APIs
# ================================

openai>=1.0.0
replicate>=0.15.0

